// src/lib/api/types.gen.ts
export type OccurrenceSummaryDtov1 = {
  /**
   * Payment Occurrence Uuid
   */
  payment_occurrence_uuid: string;
  /**
   * Due Datetime
   */
  due_datetime: string;
  status_type: PaymentOccurrenceStatusType | null;
  /**
   * Paid At
   */
  paid_at: string | null;
  /**
   * Index
   */
  index: number;
};

/**
 * PaymentCategoryType
 */
export type PaymentCategoryType =
  | "RENT"
  | "UTILITIES"
  | "LOAN"
  | "FOOD"
  | "TRANSPORT"
  | "INSURANCE"
  | "SUBSCRIPTION"
  | "SALARY"
  | "FREELANCE"
  | "OTHER";
// ...
export type PaymentOccurrenceStatusType = "PAID" | "UNPAID";

/**
 * PaymentRecurrenceType
 */
export type PaymentRecurrenceType =
  | "ONE_TIME"
  | "WEEKLY"
  | "MONTHLY"
  | "CUSTOM";

/**
 * PaymentType
 */
export type PaymentType = "EXPENSE" | "INCOME";

/**
 * PaymentWithOccurrencesDTOV1
 */
export type PaymentWithOccurrencesDtov1 = {
  /**
   * Payment Uuid
   */
  payment_uuid: string;
  /**
   * User Uuid
   */
  user_uuid: string;
  payment_type: PaymentType;
  /**
   * Title
   */
  title: string;
  /**
   * Amount
   */
  amount: number;
  category_type: PaymentCategoryType;
  /**
   * Notes
   */
  notes?: string | null;
  /**
   * Is Active
   */
  is_active: boolean;
  /**
   * Start Datetime
   */
  start_datetime: string;
  recurrence_type: PaymentRecurrenceType;
  /**
   * Interval Days
   */
  interval_days?: number | null;
  /**
   * Day Of Month Anchor
   */
  day_of_month_anchor?: number | null;
  /**
   * Total Occurrences
   */
  total_occurrences?: number | null;
  /**
   * Processed Occurrences
   */
  processed_occurrences: number;
  /**
   * Notify Week Before
   */
  notify_week_before: boolean;
  /**
   * Notify Day Before
   */
  notify_day_before: boolean;
  /**
   * Notify On Day
   */
  notify_on_day: boolean;
  /**
   * Occurrences
   */
  occurrences: Array<OccurrenceSummaryDtov1>;
};